Refactor nano evaluator test fixtures
