+++ tests/evaluation/test_nano_evaluator.py
@@ -2,34 +2,13 @@ from __future__ import annotations
 
 import importlib
 import sys
-from collections.abc import Iterator
 from types import SimpleNamespace
 from typing import Any
 
 import pytest
 
 from sentence_transformers.evaluation import NanoBEIREvaluator, NanoEvaluator, SequentialEvaluator
-
-
-class FakeDataset:
-    def __init__(self, rows: list[dict[str, Any]]) -> None:
-        self.rows = rows
-        self.column_names = list(rows[0].keys()) if rows else []
-
-    def __iter__(self) -> Iterator[dict[str, Any]]:
-        return iter(self.rows)
-
-    def __len__(self) -> int:
-        return len(self.rows)
-
-    def __getitem__(self, key: str | int) -> Any:
-        if isinstance(key, str):
-            return [row[key] for row in self.rows]
-        return self.rows[key]
-
-    def map(self, fn: Any, fn_kwargs: dict[str, Any] | None = None) -> FakeDataset:
-        kwargs = fn_kwargs or {}
-        return FakeDataset([fn(row, **kwargs) for row in self.rows])
+from tests.nano_evaluator_test_utils import build_fake_datasets_module
 
 
 class FakeInformationRetrievalEvaluator:
@@ -97,38 +76,12 @@ def dummy_model() -> Any:
 
 @pytest.fixture
 def fake_datasets_module() -> Any:
-    data: dict[tuple[str, str, str], list[dict[str, Any]]] = {}
-
-    def add_split(dataset_id: str, split_name: str) -> None:
-        data[(dataset_id, "corpus", split_name)] = [
-            {"_id": f"{split_name}-d1", "text": "Document 1"},
-            {"_id": f"{split_name}-d2", "text": "Document 2"},
-        ]
-        data[(dataset_id, "queries", split_name)] = [{"_id": f"{split_name}-q1", "text": "Query 1"}]
-        data[(dataset_id, "qrels", split_name)] = [{"query-id": f"{split_name}-q1", "corpus-id": f"{split_name}-d1"}]
-
-    for split in ["NanoMSMARCO", "NanoNQ"]:
-        add_split("sentence-transformers/NanoBEIR-en", split)
-
-    for split in ["python", "java"]:
-        add_split("example/NanoFooBar", split)
-
-    split_names: dict[tuple[str, str], list[str]] = {
-        ("sentence-transformers/NanoBEIR-en", "corpus"): ["NanoMSMARCO", "NanoNQ"],
-        ("sentence-transformers/NanoBEIR-en", "queries"): ["NanoMSMARCO", "NanoNQ"],
-        ("sentence-transformers/NanoBEIR-en", "qrels"): ["NanoMSMARCO", "NanoNQ"],
-        ("example/NanoFooBar", "corpus"): ["python", "java"],
-        ("example/NanoFooBar", "queries"): ["python", "java"],
-        ("example/NanoFooBar", "qrels"): ["python", "java"],
-    }
-
-    def load_dataset(dataset_id: str, subset: str, split: str) -> FakeDataset:
-        return FakeDataset(data[(dataset_id, subset, split)])
-
-    def get_dataset_split_names(dataset_id: str, subset: str) -> list[str]:
-        return split_names[(dataset_id, subset)]
-
-    return SimpleNamespace(load_dataset=load_dataset, get_dataset_split_names=get_dataset_split_names)
+    return build_fake_datasets_module(
+        {
+            "sentence-transformers/NanoBEIR-en": ["NanoMSMARCO", "NanoNQ"],
+            "example/NanoFooBar": ["python", "java"],
+        }
+    )
 
 
 @pytest.fixture
